refactor(objectives): separate validation from paper reconstruction
Why:
- Source-local grounding and same-paper experiment reconstruction were split across extraction and the coordinator, hiding when facts become trusted and when cross-Source binding is allowed.

What:
- Move immediate Source validation and later same-document experiment reconstruction to separate direct analysis owners, then update the coordinator, tests, and owning documentation.

Impact:
- Prompt inputs, Source ordering, scientific states, Evidence identity, progress, fallback, and public contracts remain unchanged without a compatibility or forwarding path.

Verification:
- All 120 targeted Objective analysis tests passed, and compilation, scoped lint, documentation governance, diff, and ownership checks completed successfully.
