+++ backend/application/core/objectives/analysis/README.md
@@ -14,10 +14,16 @@ Read the analysis responsibilities in real research order:
 3. `source_extraction.py` inspects each routed Source and produces transient,
    source-local `ExtractedEvidenceDraft` records. It owns text and table payload
    construction, structural table repair, deterministic table parsing, model
-   extraction, current Source grounding, and route-scoped technical failures.
-4. Within-paper experiment reconstruction, Evidence materialization, and
-   cross-paper Finding synthesis remain in their existing owners until their
-   responsibility slices move here.
+   extraction, and route-scoped technical failures.
+4. `source_validation.py` immediately checks each model-authored draft against
+   the exact Source being inspected. Unsupported results abstain; supported
+   results with incomplete variable or comparison support become descriptive
+   drafts before they can enter the next Source prompt's document state.
+5. `paper_experiment.py` runs after all routed Sources have been inspected. It
+   fills only missing scope from the same paper, binds Methods and Results only
+   through exact sample identities, and derives bounded pairwise comparisons.
+6. Evidence materialization and cross-paper Finding synthesis remain in their
+   existing owners until their responsibility slices move here.
 
 `source_screening.py` owns complete Source-unit accounting, bounded framing
 batches, prompt preflight, model/repaired/fallback dispositions, and frame
@@ -29,11 +35,22 @@ ordering, model or deterministic route decisions, selection hints, and the
 document round-robin extraction queue. It preserves Source identity and cannot
 redirect a route to another paper or Source.
 
-`source_extraction.py` owns the inspection of one exact Source at a time. A
-schema-valid model response is still only a draft: unsupported results abstain,
-unsupported comparison fields are retained only as descriptive evidence, and
-provider or irrecoverable structured-output failures remain technical failed
-drafts. Same-paper Methods/Results binding occurs only after this stage.
+`source_extraction.py` owns the inspection of one exact Source at a time. It
+passes every schema-valid model draft directly to `source_validation.py` before
+updating the accepted state supplied to the next Source prompt. Provider or
+irrecoverable structured-output failures remain technical failed drafts.
+
+`source_validation.py` owns deterministic source-local acceptance, demotion,
+or abstention. It checks the reported result, comparison labels, changed
+variables, and scientific context independently and records which field
+families each Source supports. It does not call the model or bind information
+from another Source.
+
+`paper_experiment.py` owns same-document reconstruction after Source inspection
+finishes. It may join process conditions from Methods to a result only when the
+baseline and target sample identities resolve unambiguously inside that paper.
+Missing or conflicting identities remain descriptive, and no cross-document
+binding is allowed.
 
 Technical JSON parsing, provider retries, prompt schemas, and token accounting
 support this process but do not define its scientific order.
